FILE: splits getOptionsByName and getOptionsByExpansion into separate files
