+++ libs/spirit-islander/util/src/index.ts
@@ -8,7 +8,7 @@ export * from './lib/game/difficulty';
 export * from './lib/game/expansions';
 export * from './lib/game/get-adversary-by-id';
 export * from './lib/game/get-difficulty';
-export * from './lib/game/get-options';
+export * from './lib/game/get-options-by-expansion';
 export * from './lib/game/maps';
 export * from './lib/game/option';
 export * from './lib/game/players';
