+++ libs/spirit-islander/util/src/lib/game/get-options-by-expansion.ts
@@ -1,30 +1,4 @@
 import type { ExpansionName, ExpansionOption } from './expansions';
-import type { Option } from './option';
-
-/**
- * Get options that have specified names
- * @param options
- * Array of options
- * @param names
- * Array of names used to filter returned options
- * @returns
- * Array of options that have specified names
- */
-export function getOptionsByName<
-  TName extends string,
-  TOption extends Option<TName>
->(options: readonly TOption[], names: readonly TName[]): readonly TOption[] {
-  const filteredOptions: TOption[] = [];
-
-  for (const name of names) {
-    const foundOption = options.find((option) => option.name === name);
-    if (foundOption) {
-      filteredOptions.push(foundOption);
-    }
-  }
-
-  return filteredOptions;
-}
 
 /**
  * Get options from base game and specified expansions
